restructured and added setup.py for pip installation
