+++ mtga_follower/requirements.txt
@@ -1,0 +1,2 @@
+python-dateutil
+requests
